chore(helmcharts) rename package helm/chart to helmchart
