+++ pkg/plugins/helm/changelog.go
@@ -1,4 +1,4 @@
-package helmchart
+package helm
 
 import (
 	"bytes"
