+++ pkg/plugins/helm/source.go
@@ -1,4 +1,4 @@
-package helmchart
+package helm
 
 import (
 	"fmt"
