+++ pkg/plugins/helm/target.go
@@ -1,4 +1,4 @@
-package helmchart
+package helm
 
 import (
 	"fmt"
